+++ db/migrate/20120103131914_completar_depositos.rb
@@ -1,0 +1,11 @@
+class CompletarDepositos < ActiveRecord::Migration
+  def self.up
+    require 'migrator'
+
+    Migrator.migrate(true)
+  end
+
+  def self.down
+  end
+end
+